// src/app/SelectScreen/model/types.ts
export interface CharacterProperties {
  id: string;
  name: string;
  series: string;
}

export enum CharacterColor {
  Color1 = 0,
  Color2 = 1,
  Color3 = 2,
  Color4 = 3,
  Color5 = 4,
  Color6 = 5,
  Color7 = 6,
  Color8 = 7
}

export enum CharacterImage {
  SMALL = 'small',
  HALF = 'half',
  ICON = 'icon'
}